Move KFS/TFS utils out out precision.hpp & tensorinfo.hpp
