+++ cppclean.sh
@@ -20,28 +20,36 @@ cppclean ./src/ 2>&1 | grep -v test > ${CPPCLEAN_RESULTS_FILE_SRC};
 cppclean ./src/ 2>&1 | grep test > ${CPPCLEAN_RESULTS_FILE_TEST};
 NO_WARNINGS=$(wc -l ${CPPCLEAN_RESULTS_FILE_SRC} | awk '{print $1}')
 NO_WARNINGS_TEST=$(wc -l ${CPPCLEAN_RESULTS_FILE_TEST} | awk '{print $1}')
+NO_WARNINGS_FORWARD=$(grep "use a forward declaration instead" ${CPPCLEAN_RESULTS_FILE_SRC} | wc -l)
+NO_WARNINGS_DIRECT=$(grep "not found in any directly #included header" ${CPPCLEAN_RESULTS_FILE_SRC} | wc -l)
+NO_WARNINGS_NOTUSED=$(grep " not used$" ${CPPCLEAN_RESULTS_FILE_SRC} | wc -l)
 echo "Number of warnings:" ${NO_WARNINGS}
 echo "Number of warnings in tests:" ${NO_WARNINGS_TEST}
+echo "Number of warnings about not using forward delares:" ${NO_WARNINGS_FORWARD}
+echo "Number of warnings about not direct includes:" ${NO_WARNINGS_DIRECT}
+echo "Number of warnings about not used: " ${NO_WARNINGS_NOTUSED}
 
 trap "cat ${CPPCLEAN_RESULTS_FILE_SRC}" err exit
-if [ $(grep "use a forward declaration instead" ${CPPCLEAN_RESULTS_FILE_SRC} | wc -l ) -gt 6 ]; then
-    echo "Failed due to not using forward declarations where possible";
+if [ ${NO_WARNINGS_FORWARD} -gt 6 ]; then
+    echo "Failed due to not using forward declarations where possible: ${NO_WARNINGS_FORWARD}";
     exit 1;
 fi
-if [ $(grep "not found in any directly #included header" ${CPPCLEAN_RESULTS_FILE_SRC} | wc -l ) -gt 14 ]; then
-    echo "Failed probably due to not using static keyword with functions definitions";
+if [ ${NO_WARNINGS_DIRECT} -gt 14 ]; then
+    echo "Failed probably due to not using static keyword with functions definitions: ${NO_WARNINGS_DIRECT}";
     exit 1;
 fi
-if [ $(grep " not used$" ${CPPCLEAN_RESULTS_FILE_SRC} | wc -l ) -gt 14 ]; then
-    echo "Failed probably due to unnecessary forward include";
+if [ ${NO_WARNINGS_NOTUSED} -gt 14 ]; then
+    echo "Failed probably due to unnecessary forward includes: ${NO_WARNINGS_NOTUSED}";
     exit 1;
 fi
 if [ ${NO_WARNINGS} -gt  183 ]; then
+    echo "Failed due to higher than allowed number of issues in code: ${NO_WARNINGS}"
     exit 1
 else
     exit 0;
 fi
 if [ ${NO_WARNINGS_TEST} -gt  131 ]; then
+    echo "Failed due to higher than allowed number of issues in test code: ${NO_WARNINGS_TEST}"
     exit 1
 else
     exit 0;
