+++ src/BUILD
@@ -110,8 +110,12 @@ cc_library(
         "gatherexitnodeinputhandler.hpp",
         "gcsfilesystem.cpp",
         "gcsfilesystem.hpp",
+        "grpc_utils.cpp",
+        "grpc_utils.hpp",
         "grpcservermodule.cpp",
         "grpcservermodule.hpp",
+        "kfs_frontend/kfs_utils.cpp",
+        "kfs_frontend/kfs_utils.hpp",
         "kfs_grpc_inference_service.cpp",
         "kfs_grpc_inference_service.hpp",
         "metric.cpp",
