+++ src/gatherexitnodeinputhandler.cpp
@@ -34,21 +34,4 @@ Status prepareConsolidatedTensorImpl(tensorflow::serving::PredictResponse* respo
     bufferOut = it->second.mutable_tensor_content()->data();
     return StatusCode::OK;
 }
-
-Status prepareConsolidatedTensorImpl(::inference::ModelInferResponse* response, char*& bufferOut, const std::string& name, size_t size) {
-    OVMS_PROFILE_FUNCTION();
-    for (int i = 0; i < response->outputs_size(); i++) {
-        if (response->mutable_outputs(i)->name() == name) {
-            SPDLOG_LOGGER_ERROR(dag_executor_logger, "Failed to prepare consolidated tensor, tensor with name {} already prepared", name);
-            return StatusCode::INTERNAL_ERROR;
-        }
-    }
-    auto* proto = response->add_outputs();
-    proto->set_name(name);
-    auto* content = response->add_raw_output_contents();
-    content->resize(size);
-    bufferOut = content->data();
-    return StatusCode::OK;
-}
-
 }  // namespace ovms
