+++ src/pipeline_factory.cpp
@@ -96,8 +96,8 @@ Status PipelineFactory::createInternal(std::unique_ptr<Pipeline>& pipeline,
 }
 Status PipelineFactory::create(std::unique_ptr<Pipeline>& pipeline,
     const std::string& name,
-    const ::inference::ModelInferRequest* request,
-    ::inference::ModelInferResponse* response,
+    const ::KFSRequest* request,
+    ::KFSResponse* response,
     ModelManager& manager) const {
     return this->createInternal(pipeline, name, request, response, manager);
 }
