+++ src/pipelinedefinition.cpp
@@ -1376,10 +1376,10 @@ template Status PipelineDefinition::create<tensorflow::serving::PredictRequest, 
     const tensorflow::serving::PredictRequest* request,
     tensorflow::serving::PredictResponse* response,
     ModelManager& manager);
-template Status PipelineDefinition::create<::inference::ModelInferRequest, ::inference::ModelInferResponse>(
+template Status PipelineDefinition::create<::KFSRequest, ::KFSResponse>(
     std::unique_ptr<Pipeline>& pipeline,
-    const ::inference::ModelInferRequest* request,
-    ::inference::ModelInferResponse* response,
+    const ::KFSRequest* request,
+    ::KFSResponse* response,
     ModelManager& manager);
 
 }  // namespace ovms
