+++ src/precision.hpp
@@ -19,14 +19,7 @@
 
 #include <openvino/openvino.hpp>
 
-using KFSDataType = std::string;
-namespace tensorflow {
-enum DataType : int;
-}
-
 namespace ovms {
-
-using TFSDataType = tensorflow::DataType;
 
 enum class Precision {
     BF16,
@@ -57,13 +50,7 @@ const std::string& toString(Precision precision);
 
 Precision fromString(const std::string& s);
 
-Precision KFSPrecisionToOvmsPrecision(const KFSDataType& s);
-Precision TFSPrecisionToOvmsPrecision(const TFSDataType& s);
 Precision ovElementTypeToOvmsPrecision(ov::element::Type_t type);
-
-size_t KFSDataTypeSize(const KFSDataType& datatype);
-
-const KFSDataType& ovmsPrecisionToKFSPrecision(Precision precision);
 
 ov::element::Type_t ovmsPrecisionToIE2Precision(Precision precision);
