+++ src/prediction_service.cpp
@@ -29,6 +29,7 @@
 
 #include "execution_context.hpp"
 #include "get_model_metadata_impl.hpp"
+#include "grpc_utils.hpp"
 #include "modelinstance.hpp"
 #include "modelinstanceunloadguard.hpp"
 #include "modelmanager.hpp"
@@ -110,7 +111,7 @@ grpc::Status ovms::PredictionServiceImpl::Predict(
             INCREMENT_IF_ENABLED(modelInstance->getMetricReporter().requestFailGrpcPredict);
         }
         SPDLOG_INFO("Getting modelInstance or pipeline failed. {}", status.string());
-        return status.grpc();
+        return grpc(status);
     }
 
     ExecutionContext executionContext{
@@ -126,7 +127,7 @@ grpc::Status ovms::PredictionServiceImpl::Predict(
     }
 
     if (!status.ok()) {
-        return status.grpc();
+        return grpc(status);
     }
 
     timer.stop(TOTAL);
@@ -145,7 +146,7 @@ grpc::Status PredictionServiceImpl::GetModelMetadata(
     const tensorflow::serving::GetModelMetadataRequest* request,
     tensorflow::serving::GetModelMetadataResponse* response) {
     OVMS_PROFILE_FUNCTION();
-    return getModelMetadataImpl.getModelStatus(request, response, ExecutionContext(ExecutionContext::Interface::GRPC, ExecutionContext::Method::GetModelMetadata)).grpc();
+    return grpc(getModelMetadataImpl.getModelStatus(request, response, ExecutionContext(ExecutionContext::Interface::GRPC, ExecutionContext::Method::GetModelMetadata)));
 }
 
 const GetModelMetadataImpl& PredictionServiceImpl::getTFSModelMetadataImpl() const {
