+++ src/test/kfs_rest_test.cpp
@@ -312,7 +312,7 @@ TEST_F(HttpRestApiHandlerTest, inferRequest) {
 TEST_F(HttpRestApiHandlerTest, inferPreprocess) {
     std::string request_body("{\"inputs\":[{\"name\":\"b\",\"shape\":[1,10],\"datatype\":\"FP32\",\"data\":[0,1,2,3,4,5,6,7,8,9]}],\"parameters\":{\"binary_data_output\":1, \"bool_test\":true, \"string_test\":\"test\"}}");
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request), ovms::StatusCode::OK);
 
     ASSERT_EQ(grpc_request.model_name(), modelName);
@@ -337,7 +337,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsINT8) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"INT8\",\"parameters\":{\"binary_data_size\":4}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -363,7 +363,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsBYTES) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"BYTES\",\"parameters\":{\"binary_data_size\":4}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -385,7 +385,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsINT16) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"INT16\",\"parameters\":{\"binary_data_size\":8}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -411,7 +411,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsINT32) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"INT32\",\"parameters\":{\"binary_data_size\":16}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -437,7 +437,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsINT64) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"INT64\",\"parameters\":{\"binary_data_size\":32}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -463,7 +463,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsFP32) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"FP32\",\"parameters\":{\"binary_data_size\":16}}]}";
     request_body.append((char*)values, 16);
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - 16);
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -489,7 +489,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsFP64) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"FP64\",\"parameters\":{\"binary_data_size\":32}}]}";
     request_body.append((char*)values, 32);
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - 32);
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -515,7 +515,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsBinaryDataAndContentField) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"data\":[0,1,2,3,4,5,6,7,8,9], \"datatype\":\"INT8\",\"parameters\":{\"binary_data_size\":4}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request), ovms::StatusCode::REST_CONTENTS_FIELD_NOT_EMPTY);
 }
 
@@ -524,7 +524,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsBufferSmallerThanExpected) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"INT32\",\"parameters\":{\"binary_data_size\":16}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::REST_BINARY_BUFFER_EXCEEDED);
 }
@@ -534,7 +534,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsInvalidBinaryDataSizeParameter) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[1,4],\"datatype\":\"INT32\",\"parameters\":{\"binary_data_size\":true}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::REST_BINARY_DATA_SIZE_PARAMETER_INVALID);
 }
@@ -544,7 +544,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsINT8BatchSize2) {
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[2,4],\"datatype\":\"INT8\",\"parameters\":{\"binary_data_size\":\"4, 4\"}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::OK);
 
@@ -570,7 +570,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsBinaryDataSizeStringParameterInvalid)
     std::string request_body = "{\"inputs\":[{\"name\":\"b\",\"shape\":[2,4],\"datatype\":\"INT8\",\"parameters\":{\"binary_data_size\":\"a, 4\"}}]}";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::REST_BINARY_DATA_SIZE_PARAMETER_INVALID);
 }
@@ -580,7 +580,7 @@ TEST_F(HttpRestApiHandlerTest, binaryInputsEmptyRequest) {
     std::string request_body = "";
     request_body += binaryData;
 
-    ::inference::ModelInferRequest grpc_request;
+    ::KFSRequest grpc_request;
     int inferenceHeaderContentLength = (request_body.size() - binaryData.size());
     ASSERT_EQ(HttpRestApiHandler::prepareGrpcRequest(modelName, modelVersion, request_body, grpc_request, inferenceHeaderContentLength), ovms::StatusCode::JSON_INVALID);
 }
