+++ src/test/server_test.cpp
@@ -88,8 +88,8 @@ public:
     }
     void verifyModelReady(const std::string& modelName, grpc::StatusCode expectedStatus = grpc::StatusCode::OK, bool ready = true) {
         ClientContext context;
-        ::inference::ModelReadyRequest request;
-        ::inference::ModelReadyResponse response;
+        ::KFSGetModelStatusRequest request;
+        ::KFSGetModelStatusResponse response;
         request.set_name(modelName);
 
         ASSERT_NE(nullptr, stub_);
@@ -100,8 +100,8 @@ public:
 
     void verifyServerMetadata(grpc::StatusCode expectedStatus = grpc::StatusCode::OK) {
         ClientContext context;
-        ::inference::ServerMetadataRequest request;
-        ::inference::ServerMetadataResponse response;
+        ::KFSServerMetadataRequest request;
+        ::KFSServerMetadataResponse response;
         ASSERT_NE(nullptr, stub_);
         auto status = stub_->ServerMetadata(&context, request, &response);
         ASSERT_EQ(status.error_code(), expectedStatus);
