+++ src/tfs_frontend/tfs_utils.cpp
@@ -25,56 +25,56 @@
 
 namespace ovms {
 
-tensorflow::DataType getPrecisionAsDataType(Precision precision) {
-    static std::unordered_map<Precision, tensorflow::DataType> precisionMap{
-        {Precision::FP32, tensorflow::DataType::DT_FLOAT},
-        {Precision::FP64, tensorflow::DataType::DT_DOUBLE},
-        {Precision::FP16, tensorflow::DataType::DT_HALF},
-        {Precision::I64, tensorflow::DataType::DT_INT64},
-        {Precision::I32, tensorflow::DataType::DT_INT32},
-        {Precision::I16, tensorflow::DataType::DT_INT16},
-        {Precision::I8, tensorflow::DataType::DT_INT8},
-        {Precision::U64, tensorflow::DataType::DT_UINT64},
-        {Precision::U16, tensorflow::DataType::DT_UINT16},
-        {Precision::U8, tensorflow::DataType::DT_UINT8},
-        //    {Precision::MIXED, tensorflow::DataType::DT_INVALID},
-        //    {Precision::Q78, tensorflow::DataType::DT_INVALID},
-        //    {Precision::BIN, tensorflow::DataType::DT_INVALID},
-        {Precision::BOOL, tensorflow::DataType::DT_BOOL}
-        //    {Precision::CUSTOM, tensorflow::DataType::DT_INVALID}
+TFSDataType getPrecisionAsDataType(Precision precision) {
+    static std::unordered_map<Precision, TFSDataType> precisionMap{
+        {Precision::FP32, TFSDataType::DT_FLOAT},
+        {Precision::FP64, TFSDataType::DT_DOUBLE},
+        {Precision::FP16, TFSDataType::DT_HALF},
+        {Precision::I64, TFSDataType::DT_INT64},
+        {Precision::I32, TFSDataType::DT_INT32},
+        {Precision::I16, TFSDataType::DT_INT16},
+        {Precision::I8, TFSDataType::DT_INT8},
+        {Precision::U64, TFSDataType::DT_UINT64},
+        {Precision::U16, TFSDataType::DT_UINT16},
+        {Precision::U8, TFSDataType::DT_UINT8},
+        //    {Precision::MIXED, TFSDataType::DT_INVALID},
+        //    {Precision::Q78, TFSDataType::DT_INVALID},
+        //    {Precision::BIN, TFSDataType::DT_INVALID},
+        {Precision::BOOL, TFSDataType::DT_BOOL}
+        //    {Precision::CUSTOM, TFSDataType::DT_INVALID}
     };
     auto it = precisionMap.find(precision);
     if (it == precisionMap.end()) {
-        return tensorflow::DataType::DT_INVALID;
+        return TFSDataType::DT_INVALID;
     }
     return it->second;
 }
 
-std::string getDataTypeAsString(tensorflow::DataType dataType) {
+std::string getDataTypeAsString(TFSDataType dataType) {
     switch (dataType) {
-    case tensorflow::DataType::DT_FLOAT:
+    case TFSDataType::DT_FLOAT:
         return "FP32";
-    case tensorflow::DataType::DT_DOUBLE:
+    case TFSDataType::DT_DOUBLE:
         return "FP64";
-    case tensorflow::DataType::DT_INT32:
+    case TFSDataType::DT_INT32:
         return "I32";
-    case tensorflow::DataType::DT_INT8:
+    case TFSDataType::DT_INT8:
         return "I8";
-    case tensorflow::DataType::DT_UINT8:
+    case TFSDataType::DT_UINT8:
         return "U8";
-    case tensorflow::DataType::DT_HALF:
+    case TFSDataType::DT_HALF:
         return "FP16";
-    case tensorflow::DataType::DT_INT16:
+    case TFSDataType::DT_INT16:
         return "I16";
-    case tensorflow::DataType::DT_UINT16:
+    case TFSDataType::DT_UINT16:
         return "U16";
-    case tensorflow::DataType::DT_UINT64:
+    case TFSDataType::DT_UINT64:
         return "U64";
-    case tensorflow::DataType::DT_INT64:
+    case TFSDataType::DT_INT64:
         return "I64";
-    case tensorflow::DataType::DT_BOOL:
+    case TFSDataType::DT_BOOL:
         return "BOOL";
-    case tensorflow::DataType::DT_STRING:
+    case TFSDataType::DT_STRING:
         return "STRING";
     default:
         return "INVALID";
@@ -95,4 +95,24 @@ std::string tensorShapeToString(const tensorflow::TensorShapeProto& tensorShape)
 
     return oss.str();
 }
+Precision TFSPrecisionToOvmsPrecision(const TFSDataType& datatype) {
+    static std::unordered_map<TFSDataType, Precision> precisionMap{
+        {TFSDataType::DT_FLOAT, Precision::FP32},
+        {TFSDataType::DT_DOUBLE, Precision::FP64},
+        {TFSDataType::DT_HALF, Precision::FP16},
+        {TFSDataType::DT_INT64, Precision::I64},
+        {TFSDataType::DT_INT32, Precision::I32},
+        {TFSDataType::DT_INT16, Precision::I16},
+        {TFSDataType::DT_INT8, Precision::I8},
+        {TFSDataType::DT_UINT64, Precision::U64},
+        {TFSDataType::DT_UINT16, Precision::U16},
+        {TFSDataType::DT_UINT8, Precision::U8},
+        {TFSDataType::DT_BOOL, Precision::BOOL}};
+    auto it = precisionMap.find(datatype);
+    if (it == precisionMap.end()) {
+        return Precision::UNDEFINED;
+    }
+    return it->second;
+}
+
 }  // namespace ovms
